src/com/mycompany/almacenapp/AlmacenTest.java
package com.mycompany.almacenapp;

import static org.junit.jupiter.api.Assertions.*;
import org.junit.jupiter.api.*;

import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;
import java.sql.Statement;
import java.util.ArrayList;
import java.util.List;

class AlmacenTest {

    private AlmacenTest almacen;
    private Connection conn;
    private Statement stmt;

    @BeforeEach
    void setUp() throws SQLException {
        // Establecer una conexión a la base de datos SQLite para pruebas
        conn = DriverManager.getConnection("jdbc:sqlite:inventario_test.db");
        stmt = conn.createStatement();

        // Crear tablas necesarias para las pruebas
        String createInventarioTable = "CREATE TABLE IF NOT EXISTS inventario (" +
                "id INTEGER PRIMARY KEY AUTOINCREMENT," +
                "producto TEXT," +
                "peso REAL," +
                "ubicacion TEXT," +
                "fecha TEXT," +
                "tipo TEXT)";
        stmt.execute(createInventarioTable);

        String createUsuariosTable = "CREATE TABLE IF NOT EXISTS usuarios (" +
                "id INTEGER PRIMARY KEY AUTOINCREMENT," +
                "username TEXT UNIQUE," +
                "password TEXT," +
                "rol TEXT)";
        stmt.execute(createUsuariosTable);

        // Inicializar la instancia de Almacen
        almacen = new AlmacenTest();
    }

    @AfterEach
    void tearDown() throws SQLException {
        // Cerrar la conexión a la base de datos después de cada prueba
        stmt.execute("DROP TABLE IF EXISTS inventario");
        stmt.execute("DROP TABLE IF EXISTS usuarios");
        stmt.close();
        conn.close();
    }

    @Test
    void testCrearUsuario() {
        almacen.crearUsuario("testuser", "password", "admin");
        String rol = almacen.verificarUsuario("testuser", "password");
        assertEquals("admin", rol);
    }

    @Test
    void testEntradaProducto() {
        almacen.entradaProducto("TestProduct", new ArrayList<>(List.of(5.0, 10.0)), "Planta 1 - Almacén Central");
        // Aquí podrías agregar más lógica para verificar que los productos se hayan agregado correctamente
    }

    @Test
    void testSalidaProducto() {
        almacen.entradaProducto("TestProduct", new ArrayList<>(List.of(5.0, 10.0)), "Planta 1 - Almacén Central");
        almacen.salidaProducto("TestProduct", new ArrayList<>(List.of(5.0)), "Planta 1 - Almacén Central");
        // Aquí podrías agregar más lógica para verificar que los productos se hayan eliminado correctamente
    }

    @Test
    void testVerificarUsuario() {
        almacen.crearUsuario("testuser", "password", "admin");
        String rol = almacen.verificarUsuario("testuser", "password");
        assertEquals("admin", rol);

    private void crearUsuario(String testuser, String password, String admin) {
        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
    }

    private String verificarUsuario(String testuser, String password) {
        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
    }

    private void assertEquals(String admin, String rol) {
        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
    }

    private void entradaProducto(String testProduct, ArrayList<Double> arrayList, String planta_1__Almacén_Central) {
        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
    }

    private void salidaProducto(String testProduct, ArrayList<Double> arrayList, String planta_1__Almacén_Central) {
        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
   }
